Tidy up metrics and their names
This removes the fahrenheit and kelvin metrics plus the ones reporting in
miles. Based on advice from https://prometheus.io/docs/practices/naming/.

It should be up to the presentation of the metrics to convert the values from
one unit to another, especially as it's just a matter of applying a bit of
maths.

The sunrise/sunset metrics got "_timestamp" appended and the wind direction
metric got "_degrees" appended to the metric name.

The exporter now requests the metrics from OpenWeatherMap in metric format
which means we no longer need to do any unit conversion.